+++ nifi-extension-bundles/nifi-standard-bundle/nifi-standard-processors/src/test/java/org/apache/nifi/processors/standard/PutDatabaseRecordIT.java
@@ -52,7 +52,7 @@ public class PutDatabaseRecordIT {
     private static final long MILLIS_TIMESTAMP_LONG = 1707238288351L;
     private static final long MICROS_TIMESTAMP_LONG = 1707238288351567L;
     private static final String MICROS_TIMESTAMP_FORMATTED = "2024-02-06 11:51:28.351567";
-    private static final double MICROS_TIMESTAMP_DOUBLE = ((double) MICROS_TIMESTAMP_LONG) / 1000000D;
+    private static final double MICROS_TIMESTAMP_DOUBLE = (MICROS_TIMESTAMP_LONG) / 1000000D;
     private static final long NANOS_AFTER_SECOND = 351567000L;
     private static final Instant INSTANT_MICROS_PRECISION = Instant.ofEpochMilli(MILLIS_TIMESTAMP_LONG).plusNanos(NANOS_AFTER_SECOND).minusMillis(MILLIS_TIMESTAMP_LONG % 1000);
