+++ nifi-extension-bundles/nifi-standard-bundle/nifi-standard-reporting-tasks/src/main/java/org/apache/nifi/controller/MonitorMemory.java
@@ -212,7 +212,7 @@ public class MonitorMemory extends AbstractReportingTask {
             return;
         }
 
-        final double percentageUsed = (double) usage.getUsed() / (double) usage.getMax() * 100D;
+        final double percentageUsed = (double) usage.getUsed() / usage.getMax() * 100D;
         // In certain scenarios in the monitored memory bean the gcSensor can get stuck in 'on' state before the usage would reach the threshold
         // and this will cause false exceeded state until the next garbage collection. To eliminate this we are adding a condition with the calculated usage threshold.
         if (bean.isCollectionUsageThresholdSupported() && bean.isCollectionUsageThresholdExceeded() && usage.getUsed() > calculatedThreshold) {
